Rafactored DataRegister
change name
- data_register to seed_data_register

add files
- seed_data_register,
- registration_config_updater
