+++ lib/pokotarou/seeder.rb
@@ -12,7 +12,6 @@ class Seeder
       return make_array(n, ->(){ SecureRandom.hex(20) }) if type == "string"
       return make_array(n, ->(){ SecureRandom.hex(200) }) if ["text", "binary"].include?(type)
       return make_array(n, ->(){ [true, false].sample }) if type == "boolean"
-      return make_string_array(n, enum) if type == "string"
       return make_datetime_array() if ["datetime", "date", "time"].include?(type)
     end
