+++ test/data/methods.rb
@@ -24,4 +24,4 @@ end
 
 def alphabet_array 
   ["A", "B", "C"]
-end
+end
